add new unit tests
