+++ met-web/tests/unit/components/engagement/form/edit/EngagementForm.Edit.test.tsx
@@ -3,7 +3,6 @@ import React from 'react';
 import '@testing-library/jest-dom';
 import EngagementForm from '../../../../../../src/components/engagement/form';
 import { setupEnv } from '../../../setEnvVars';
-import * as reactRedux from 'react-redux';
 import * as reactRouter from 'react-router';
 import * as engagementService from 'services/engagementService';
 import * as engagementMetadataService from 'services/engagementMetadataService';
@@ -12,10 +11,8 @@ import * as notificationModalSlice from 'services/notificationModalService/notif
 import * as widgetService from 'services/widgetService';
 import { createDefaultSurvey, Survey } from 'models/survey';
 import { WidgetType } from 'models/widget';
-import { Box } from '@mui/material';
 import { draftEngagement, engagementMetadata, engagementSlugData } from '../../../factory';
-import { USER_ROLES } from 'services/userService/constants';
-import assert from 'assert';
+import { setupCommonMocks } from './jestTestUtils';
 
 const survey: Survey = {
     ...createDefaultSurvey(),
@@ -26,55 +23,10 @@ const survey: Survey = {
 
 const surveys = [survey];
 
-jest.mock('axios')
-
-jest.mock('react-redux', () => ({
-    ...jest.requireActual('react-redux'),
-    useSelector: jest.fn(() => {
-        return {
-            roles: [USER_ROLES.VIEW_PRIVATE_ENGAGEMENTS, USER_ROLES.EDIT_ENGAGEMENT, USER_ROLES.CREATE_ENGAGEMENT],
-            assignedEngagements: [draftEngagement.id],
-        };
-    }),
-}));
-
-jest.mock('components/common/Dragdrop', () => ({
-    ...jest.requireActual('components/common/Dragdrop'),
-    MetDroppable: ({ children }: { children: React.ReactNode }) => <Box>{children}</Box>,
-    MetDraggable: ({ children }: { children: React.ReactNode }) => <Box>{children}</Box>,
-}));
-
-jest.mock('@hello-pangea/dnd', () => ({
-    ...jest.requireActual('@hello-pangea/dnd'),
-    DragDropContext: ({ children }: { children: React.ReactNode }) => <Box>{children}</Box>,
-}));
-
-jest.mock('@reduxjs/toolkit/query/react', () => ({
-    ...jest.requireActual('@reduxjs/toolkit/query/react'),
-    fetchBaseQuery: jest.fn(),
-}));
-
-jest.mock('components/map', () => () => {
-    return <Box></Box>;
-});
-
-jest.mock('apiManager/apiSlices/widgets', () => ({
-    ...jest.requireActual('apiManager/apiSlices/widgets'),
-    useCreateWidgetMutation: () => [jest.fn(() => Promise.resolve())],
-    useDeleteWidgetMutation: () => [jest.fn(() => Promise.resolve())],
-    useSortWidgetsMutation: () => [jest.fn(() => Promise.resolve())],
-}));
-
-// Mocking window.location.pathname in Jest
-Object.defineProperty(window, 'location', {
-    value: {
-        pathname: '/engagements/1/form',
-    },
-});
-
 describe('Engagement form page tests', () => {
-    jest.spyOn(reactRedux, 'useDispatch').mockImplementation(() => jest.fn());
-    jest.spyOn(reactRouter, 'useNavigate').mockImplementation(() => jest.fn());
+    beforeAll(() => {
+        setupCommonMocks();
+    });
     const openNotificationModalMock = jest
         .spyOn(notificationModalSlice, 'openNotificationModal')
         .mockImplementation(jest.fn());
@@ -82,75 +34,16 @@ describe('Engagement form page tests', () => {
     const getEngagementMetadataMock = jest
         .spyOn(engagementMetadataService, 'getEngagementMetadata')
         .mockReturnValue(Promise.resolve(engagementMetadata));
-    jest.spyOn(engagementMetadataService, 'patchEngagementMetadata').mockReturnValue(
-        Promise.resolve(engagementMetadata),
-    );
     const getEngagementSlugMock = jest
         .spyOn(engagementSlugService, 'getSlugByEngagementId')
         .mockReturnValue(Promise.resolve(engagementSlugData));
     const getEngagementMock = jest
         .spyOn(engagementService, 'getEngagement')
         .mockReturnValue(Promise.resolve(draftEngagement));
-    const patchEngagementMock = jest
-        .spyOn(engagementService, 'patchEngagement')
-        .mockReturnValue(Promise.resolve(draftEngagement));
     const getWidgetsMock = jest.spyOn(widgetService, 'getWidgets').mockReturnValue(Promise.resolve([]));
 
     beforeEach(() => {
         setupEnv();
-    });
-
-    test('Engagement form with saved engagement should display saved info', async () => {
-        useParamsMock.mockReturnValue({ engagementId: '1' });
-        const { container } = render(<EngagementForm />);
-
-        await waitFor(() => {
-            expect(screen.getByDisplayValue('Test Engagement')).toBeInTheDocument();
-            expect(container.querySelector('span.MuiSkeleton-root')).toBeNull();
-        });
-
-        expect(getEngagementMock).toHaveBeenCalledOnce();
-        expect(getEngagementMetadataMock).toHaveBeenCalledOnce();
-        expect(screen.getByTestId('update-engagement-button')).toBeVisible();
-        expect(screen.getByDisplayValue('2022-09-01')).toBeInTheDocument();
-        expect(screen.getByDisplayValue('2022-09-30')).toBeInTheDocument();
-        expect(screen.getByText('Survey 1')).toBeInTheDocument();
-    });
-
-    test('Save engagement button should trigger patch call', async () => {
-        useParamsMock.mockReturnValue({ engagementId: '1' });
-        const { container } = render(<EngagementForm />);
-
-        await waitFor(() => {
-            expect(screen.getByDisplayValue('Test Engagement')).toBeInTheDocument();
-            expect(container.querySelector('span.MuiSkeleton-root')).toBeNull();
-        });
-        const updateButton = screen.getByTestId('update-engagement-button');
-
-        const nameInput = container.querySelector('input[name="name"]');
-        assert(nameInput, 'Unable to find engagement name input');
-        fireEvent.change(nameInput, { target: { value: 'Survey 1 Updated' } });
-        fireEvent.click(updateButton);
-
-        await waitFor(() => {
-            expect(patchEngagementMock).toHaveBeenCalledOnce();
-        });
-    });
-
-    test('Modal with warning appears when removing survey', async () => {
-        useParamsMock.mockReturnValue({ engagementId: '1' });
-        const { container } = render(<EngagementForm />);
-
-        await waitFor(() => {
-            expect(screen.getByDisplayValue('Test Engagement')).toBeInTheDocument();
-            expect(container.querySelector('span.MuiSkeleton-root')).toBeNull();
-        });
-
-        const removeSurveyButton = screen.getByTestId(`survey-widget/remove-${survey.id}`);
-
-        fireEvent.click(removeSurveyButton);
-
-        expect(openNotificationModalMock).toHaveBeenCalledOnce();
     });
 
     test('Cannot add more than one survey', async () => {
